fix(nix): Build rplidar-ros from source for Jazzy
rplidar-ros is not officially released for ROS 2 Jazzy in the
rosdistro repository. Previously, patches.nix attempted to override
rosPrev.rplidar-ros, but this package doesn't exist in the Jazzy
distribution of nix-ros-overlay.

This fixes the "package not found" error that occurs when updating
nix-ros-overlay to newer versions where rplidar-ros is not available
for Jazzy.